Root A and AAAA records return properly for mixed-case queries
When queried with a mixed-case domain, e.g. "SsLiP.Io", the A and AAAA
records are properly returned. Previously they weren't. This is a
problem when Let's Encrypt queries with a mixed-case domain, which it
does for reasons which are not clear:

  <https://github.com/letsencrypt/boulder/issues/1243>

Updated tests to reflect new sslip.io A and SOA records (everything is
on Hetzner now).

fixes:
```
Failed authorization procedure. sslip.io (http-01): urn:acme:error:unknownHost :: The server could not resolve a domain name :: No valid IP addresses found for sslip.io
```