`squash_rttm.py` will help test the IBM diarizer
The IBM Diarizer doesn't collapse the utterances of an individual
speaker. For example, if a speaker drones one for a minute, the
resulting RTTM file could have, say, eighty lines. But they would be
much easier to represent if they were only one line. This filter
fixes that problem, collapses the lines. It also only prints 2
decimal places.

The tests are contained within the script. I want testing, but I
loath the plethora of directories and files. Personal taste.