Install AWS sslip.io DNS server
We're not going down the k8s worker route. It'd require a load balancer,
and I'm not interested in a load balancer because it's too expensive
and, in my humble opinion, not worth it.

Instead I'm going to create a regular VM which runs sslip.io as a
service. The old-school way.