🐞 Brewfile: don't even try to install `yq`
Too much of our tooling depends on `yq` v3, but Homebrew is being
draconian about pushing us to v4 because reasons. This is a classic case
of hubris, where a developer sacrifices backwards compatibility on the
altar of aesthetic purity.

fixes:
```
Error: The `brew link` step did not complete successfully
The formula built, but is not symlinked into /usr/local
Could not symlink bin/yq
Target /usr/local/bin/yq
```
and
```
Error: yq@3 has been disabled because it is not maintained upstream!
Installing yq@3 has failed!
```