Install ns-azure: script is run twice
The script is intended to run twice, once by the cloud-init's adminuser,
and once by cunnie.

Yes, it's suboptimal to run the configure script twice, but the
alternative was a very-tortured script (for example, deleting the
adminuser while the script is being run by the adminuser opens up a
Pandora's box of things that could go wrong).

- bump chruby 0.7.0 → 0.8.3
- delete the adminuser if run by cunnie
- temporarily comment-out getting TLS certs to avoid running afoul of
  quota limits