ns-aws runs on Ubuntu, not Fedora
I want Fedora to work, but the straw that broke the camel's back is the
etcd version was too new (3.5, which etcd.io recommends downgrading to
3.4), and didn't play well with ns-azure's (Ubuntu) version (3.3).

Another problem was that Golang was stuck at 1.16; I needed at least
1.17, and I don't want to do a custom install of Golang.

It was always tricky finding a Fedora image. On Azure I couldn't find
one, so I went with Ubuntu.

I'd like to have my servers be consistent. Maintaining two
very-different configuration scripts for ns-aws & ns-azure was a
rewarding challenge, but my time is limited, and it can be more
fruitfully working on other challenges.

I stuck with Fedora a long time because my friend Mike Tiemann was
briefly the CTO, but that was over twenty years ago. My emotional
attachment to Fedora is irrational at this point.