NS-GCE: delete the weird users
I don't know why these users are added to my VMs, and I vaguely remember
adding Evan Brown to my project 8 years ago, but I can't find him in
console.cloud.google.com, and I can't be bothered to open a ticket.

Drive-by: TLS is a bust on GCE, and I think it's because "hostnames must
not start or end with a hyphen", which is not the case for the IPv6
hostname.